requirements.txt
fastapi
uvicorn
langchain
langchain-community
sentence-transformers
faiss-cpu
pypdf
rank-bm25
groq
python-dotenv
httpx==0.27.0